LF-4093 Move exported enums into own file; use enums when destructuring filter components in inventory
