+++ packages/webapp/src/containers/Filter/Animals/types.ts
@@ -1,0 +1,28 @@
+/*
+ *  Copyright 2024 LiteFarm.org
+ *  This file is part of LiteFarm.
+ *
+ *  LiteFarm is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  LiteFarm is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
+ */
+
+export enum AnimalsFilterKeys {
+  ANIMAL_OR_BATCH = 'ANIMAL_OR_BATCH',
+  TYPE = 'TYPE',
+  BREED = 'BREED',
+  SEX = 'SEX',
+  GROUPS = 'GROUPS',
+  LOCATION = 'LOCATION',
+}
+
+export enum AnimalOrBatchKeys {
+  ANIMAL = 'ANIMAL',
+  BATCH = 'BATCH',
+}
